# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OTel Data API",
    "description": "REST API for OwnTracks/Garmin GPS data with PostGIS spatial queries",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/garmin/activities": {
      "get": {
        "tags": [
          "Garmin"
        ],
        "summary": "List Activities",
        "description": "List Garmin activities with filtering and pagination.\n\nReturns paginated Garmin cycling/running activities with track point counts.\nFilter by sport type or date range. Sort by start time, distance, duration, etc.",
        "operationId": "list_activities_api_v1_garmin_activities_get",
        "parameters": [
          {
            "name": "sport",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by sport type",
              "examples": [
                "cycling"
              ],
              "title": "Sport"
            },
            "description": "Filter by sport type"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter from date (YYYY-MM-DD)",
              "examples": [
                "2025-11-01"
              ],
              "title": "Date From"
            },
            "description": "Filter from date (YYYY-MM-DD)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to date (YYYY-MM-DD)",
              "examples": [
                "2025-11-30"
              ],
              "title": "Date To"
            },
            "description": "Filter to date (YYYY-MM-DD)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort column (start_time, distance_km, duration_seconds, sport, created_at)",
              "default": "start_time",
              "title": "Sort"
            },
            "description": "Sort column (start_time, distance_km, duration_seconds, sport, created_at)"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_GarminActivity_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/garmin/activities/{activity_id}": {
      "get": {
        "tags": [
          "Garmin"
        ],
        "summary": "Get Activity",
        "description": "Get a single Garmin activity by ID with track point count.",
        "operationId": "get_activity_api_v1_garmin_activities__activity_id__get",
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Garmin activity ID",
              "examples": [
                "20932993811"
              ],
              "title": "Activity Id"
            },
            "description": "Garmin activity ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GarminActivity"
                }
              }
            }
          },
          "404": {
            "description": "Activity not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/garmin/activities/{activity_id}/tracks": {
      "get": {
        "tags": [
          "Garmin"
        ],
        "summary": "List Track Points",
        "description": "List track points for a specific activity.\n\nReturns GPS track points with altitude, speed, heart rate, and cadence data.\nSupports up to 10,000 points per request for full activity visualization.",
        "operationId": "list_track_points_api_v1_garmin_activities__activity_id__tracks_get",
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Garmin activity ID",
              "examples": [
                "20932993811"
              ],
              "title": "Activity Id"
            },
            "description": "Garmin activity ID"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort column (timestamp, altitude, speed_kmh, heart_rate, created_at)",
              "default": "timestamp",
              "title": "Sort"
            },
            "description": "Sort column (timestamp, altitude, speed_kmh, heart_rate, created_at)"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_GarminTrackPoint_"
                }
              }
            }
          },
          "404": {
            "description": "Activity not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GarminActivity": {
        "properties": {
          "activity_id": {
            "type": "string",
            "title": "Activity Id"
          },
          "sport": {
            "type": "string",
            "title": "Sport"
          },
          "sub_sport": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Sport"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "avg_heart_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Heart Rate"
          },
          "max_heart_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Heart Rate"
          },
          "avg_cadence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Cadence"
          },
          "max_cadence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Cadence"
          },
          "calories": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories"
          },
          "avg_speed_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Speed Kmh"
          },
          "max_speed_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Speed Kmh"
          },
          "total_ascent_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Ascent M"
          },
          "total_descent_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Descent M"
          },
          "total_distance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Distance"
          },
          "avg_pace": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Pace"
          },
          "device_manufacturer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Manufacturer"
          },
          "avg_temperature_c": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Temperature C"
          },
          "min_temperature_c": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Temperature C"
          },
          "max_temperature_c": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Temperature C"
          },
          "total_elapsed_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Elapsed Time"
          },
          "total_timer_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Timer Time"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "uploaded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded At"
          },
          "track_point_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Point Count"
          }
        },
        "type": "object",
        "required": [
          "activity_id",
          "sport"
        ],
        "title": "GarminActivity",
        "examples": [
          {
            "activity_id": "20932993811",
            "avg_cadence": 78,
            "avg_heart_rate": 142,
            "avg_pace": 3.513345,
            "avg_speed_kmh": 26.3,
            "avg_temperature_c": 18,
            "calories": 1689,
            "created_at": "2026-02-09T23:56:37Z",
            "device_manufacturer": "garmin",
            "distance_km": 50.6,
            "duration_seconds": 6932,
            "end_time": "2025-11-08T20:16:45Z",
            "max_cadence": 112,
            "max_heart_rate": 178,
            "max_speed_kmh": 48.7,
            "max_temperature_c": 22,
            "min_temperature_c": 14,
            "sport": "cycling",
            "start_time": "2025-11-08T18:21:13Z",
            "sub_sport": "road",
            "total_ascent_m": 385,
            "total_descent_m": 380,
            "total_distance": 50598.95,
            "total_elapsed_time": 7200.5,
            "total_timer_time": 6932.1,
            "track_point_count": 10707,
            "uploaded_at": "2026-02-09T23:56:37Z"
          }
        ]
      },
      "GarminTrackPoint": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "activity_id": {
            "type": "string",
            "title": "Activity Id"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "altitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude"
          },
          "distance_from_start_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance From Start Km"
          },
          "speed_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Kmh"
          },
          "heart_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate"
          },
          "cadence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cadence"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "activity_id",
          "latitude",
          "longitude",
          "timestamp"
        ],
        "title": "GarminTrackPoint",
        "examples": [
          {
            "activity_id": "20932993811",
            "altitude": 12.4,
            "cadence": 80,
            "created_at": "2026-02-09T23:56:37Z",
            "distance_from_start_km": 0.0,
            "heart_rate": 135,
            "id": 7454,
            "latitude": 40.71501586586237,
            "longitude": -74.01768794283271,
            "speed_kmh": 24.5,
            "temperature_c": 18,
            "timestamp": "2025-11-08T18:21:13Z"
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_GarminActivity_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GarminActivity"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[GarminActivity]"
      },
      "PaginatedResponse_GarminTrackPoint_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GarminTrackPoint"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[GarminTrackPoint]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}